validation

+++ src/main/java/com/example/tarea3_3ventas/dao/ComercialDAOImpl.java
@@ -48,7 +48,7 @@ public class ComercialDAOImpl implements ComercialDAO {
                         rs.getString("nombre"),
                         rs.getString("apellido1"),
                         rs.getString("apellido2"),
-                        rs.getFloat("comision")
+                        rs.getBigDecimal("comision")
 
                 )
         );
@@ -66,7 +66,7 @@ public class ComercialDAOImpl implements ComercialDAO {
                                 rs.getString("nombre"),
                                 rs.getString("apellido1"),
                                 rs.getString("apellido2"),
-                                rs.getFloat("comision"))
+                                rs.getBigDecimal("comision"))
                         , id
                 );
